Recognize absolute paths on Windows
Importing from an absolute path like '/path/to/file.scss' shall work
regardless of the underlying platform. But on Windows,
pathlib.is_absolute() does not return True for this string because of
missing drive symbol. This way handle_import() could not do the correct
thing. The fix is simply to treat such paths as absolute ones.